Split config docs.
