+++ .moon/project.yml
@@ -1,4 +1,4 @@
-$schema: '../schemas/global-project.json'
+$schema: '../website/static/schemas/global-project.json'
 
 fileGroups:
   configs:
