+++ packages/types/src/index.ts
@@ -5,3 +5,4 @@ export * from './project-config';
 export * from './runner';
 export * from './template-config';
 export * from './workspace-config';
+export * from './toolchain-config';
