+++ website/docs/config/global-project.mdx
@@ -17,7 +17,7 @@ Defines an external `.moon/project.yml` to extend and inherit settings from. Per
 and sharing configuration across repositories and projects. When defined, this setting must be an
 HTTPS URL _or_ relative file system path that points to a valid YAML document!
 
-```yaml title=".moon/workspace.yml" {1}
+```yaml title=".moon/project.yml" {1}
 extends: 'https://raw.githubusercontent.com/organization/repository/master/.moon/project.yml'
 ```
